further refactoring of K8sManifests
