+++ k8s-plugin/src/main/kotlin/com/amazon/spinnaker/keel/k8s/model/CredentialsResourceSpec.kt
@@ -1,7 +1,5 @@
 package com.amazon.spinnaker.keel.k8s.model
 
-import com.amazon.spinnaker.keel.k8s.K8sCredentialManifest
-import com.amazon.spinnaker.keel.k8s.K8sObjectManifest
 import com.amazon.spinnaker.keel.k8s.SECRET
 import com.amazon.spinnaker.keel.k8s.SECRET_API_V1
 import com.netflix.spinnaker.keel.api.SimpleLocations
